Build Data frame through function
In the previous commit, I didn't put the codes building dictionary to functions. In order to get structured data from different years, it's easier to deal with functions. However, the issue here is that as these functions keep calling the previous ones when running, it's actually not that effectively than code in the previous commit. I tried to yield data on 30 pages, which is about 600 movies, but the dataframe didn't come out even after one and half hour.