More web scraping and Data structured
In this file, I added the "release date" attribute of each movies to the data set, and again used regex to clean up the text from web. Finally in this file, a relatively complete loop for web scraping is done and I try to get movies from 2 pages in this file.