Dataframes of movie for each year
I tried to build the data frames based on the function built previously, however, it takes too long to run the function as it's calling other two functions inside (I quit the function after running it for about two hours for just 1 year). Therefore, I yielded the data frame in the most simple but really repetitive way, which is too repeat the code again and again manually.