Modified Web Scraping
In the previous files, I used index to get text for information with the same <p> in the front on the web page. However, I find that the structure of the page for different movies are not identical, which made a loss in getting enough information. Therefore I now modified the web scraping function, instead of indexing, I used a for loop to get information needed with specified keyword included in the text string.